requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3.1
python-decouple==3.8
requests==2.31.0
